mailinfo.c::convert_to_utf8: reuse strlen info
We're passing buffer from strbuf to reencode_string, which will
call strlen(3) on that buffer, and discard the length of newly created
buffer. Then, we compute the length of the return buffer to attach to
a strbuf.

We can do better than that by reusing all the available information
and call the underlying lower level function that will be called
indirectly by reencode_string.

Signed-off-by: Đoàn Trần Công Danh <congdanhqx@gmail.com>