mailinfo: try to recover space for marking context
Some MUA mangles the space at begining of lines to non-breaking-space,
thus renders the diff included with email invalid.

If the diff context is an empty line, we'll have a context with a single
space. In format=flowed email, that space will become a space in the end
of line, and we'll lose it. Hence, we can conclude that there's must be
at least a space that was lost in an empty line in the diff.

Recover those spaces.

Signed-off-by: Đoàn Trần Công Danh <congdanhqx@gmail.com>