fixup! CI: limit GitHub Actions to designated branches
Here is the patch to for discussion.

Should we want to enable GitHub Action for all repo-s, except git/git
and gitster/git. We'll want to have this patch fix-up to the previous
one.

With this patch merged, contributors will need to opt-out by push to
a branch/tag with "wip" anywhere in the refname.

Note that, integration branches (maint, master, next, jch, pu) will
always be built, regardless of repo.
Since the current condition is ugly enough, and I find it's very hard to
wrap my head around those condition.

Signed-off-by: Đoàn Trần Công Danh <congdanhqx@gmail.com>