ci: GitHub Action: run static analysis on ubuntu-16.04
In a later patch, we will enable sparse on our CI systems.

Since sparse should be run in the system that have the same compiler
version that compiled sparse [1]
We'd like to pull sparse from an Azure pipeline to save time.
And the Azure pipeline in question and our Travis job is running in
Ubuntu 16.04, as of now.

Limit the StaticAnalysis job on GitHub Action to ubuntu-16.04

1: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=505177#15

Signed-off-by: Đoàn Trần Công Danh <congdanhqx@gmail.com>